+++ sdk/keyvault/keyvault-keys/CHANGELOG.md
@@ -4,6 +4,7 @@
 - `deleteKey` and `recoverDeletedKey` are now out of the public API.
   Use `beginDeleteKey` and `beginRecoverDeletedKey` instead.
   They both return a Poller (from our package `@azure/core-lro`) that manages the long running operation.
+- Renamed `Key` to `KeyVaultKey`.
 
 ## 4.0.0-preview.8 (2019-10-09)